+++ xb_social_ai_planner/wizard/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 from . import xb_social_generate_month_wizard
+from . import xb_social_onboarding
 from . import xb_social_refine_wizard